API/requirements.txt
flask
tensorflow
Pillow
IPython
scipy
opencv-python
matplotlib
flask-cors
